Center camera to windows center
Map paint almost finished
TODO fix zoom when updating window size

+++ info3.game.given.2021/src/info3/game/map/ITile.java
@@ -3,5 +3,5 @@ package info3.game.map;
 import java.awt.Graphics;
 
 public interface ITile {
-	public void paint(Graphics g,int screenPosX,int screenPosY,int size);
+	public void paint(Graphics g,float screenPosX,float screenPosY,int size);
 }
